requirements.txt
Flask==2.2.2
tensorflow-cpu==2.5.0
gunicorn==20.1.0